api-main/requirements.txt
fastapi
uvicorn
opencv-python
face_recognition
pillow
numpy
